Extract a common Snapshotter interface
I couldn't copy-paste this stuff a third time.
