+++ gradle/libs.versions.toml
@@ -6,6 +6,7 @@ androidx-activity = "1.9.2"
 androidx-compose-ui = "1.7.2"
 jbCompose = "1.6.11"
 lint = "31.6.1"
+paparazzi = "1.3.2"
 zipline = "1.17.0"
 coil = "3.0.0-alpha10"
 okio = "3.9.1"
@@ -78,7 +79,8 @@ paging-compose-common = { module = "app.cash.paging:paging-compose-common", vers
 zipline = { module = "app.cash.zipline:zipline", version.ref = "zipline" }
 zipline-gradlePlugin = { module = "app.cash.zipline:zipline-gradle-plugin", version.ref = "zipline" }
 zipline-loader = { module = "app.cash.zipline:zipline-loader", version.ref = "zipline" }
-paparazzi-gradlePlugin = "app.cash.paparazzi:paparazzi-gradle-plugin:1.3.2"
+paparazzi-gradlePlugin = { module = "app.cash.paparazzi:paparazzi-gradle-plugin", version.ref = "paparazzi" }
+paparazzi = { module = "app.cash.paparazzi:paparazzi", version.ref = "paparazzi" }
 jimfs = "com.google.jimfs:jimfs:1.3.0"
 coil-compose = { module = "io.coil-kt.coil3:coil-compose-core", version.ref = "coil" }
 coil-core = { module = "io.coil-kt.coil3:coil", version.ref = "coil" }
