+++ redwood-layout-uiview/src/commonTest/kotlin/app/cash/redwood/layout/uiview/UIViewSpacerTest.kt
@@ -16,6 +16,8 @@
 package app.cash.redwood.layout.uiview
 
 import app.cash.redwood.layout.AbstractSpacerTest
+import app.cash.redwood.snapshot.testing.UIViewSnapshotCallback
+import app.cash.redwood.snapshot.testing.UIViewSnapshotter
 import app.cash.redwood.widget.Widget
 import kotlinx.cinterop.readValue
 import kotlinx.cinterop.useContents
