+++ redwood-snapshot-testing/build.gradle
@@ -1,0 +1,28 @@
+import static app.cash.redwood.buildsupport.TargetGroup.CommonWithAndroid
+
+apply plugin: 'org.jetbrains.kotlin.plugin.compose'
+
+redwoodBuild {
+  targets(CommonWithAndroid)
+}
+
+kotlin {
+  sourceSets {
+    commonMain {
+      dependencies {
+        api libs.jetbrains.compose.foundation
+        api libs.kotlin.test
+        implementation libs.assertk
+      }
+    }
+    androidMain {
+      dependencies {
+        api libs.paparazzi
+      }
+    }
+  }
+}
+
+android {
+  namespace 'app.cash.redwood.snapshot.testing'
+}
