+++ settings.gradle
@@ -75,6 +75,7 @@ include ':redwood-protocol-guest'
 include ':redwood-protocol-host'
 include ':redwood-runtime'
 include ':redwood-schema'
+include ':redwood-snapshot-testing'
 include ':redwood-testing'
 include ':redwood-tooling-codegen'
 include ':redwood-tooling-lint'
